chore(ante) add ante authz decorator test
